Add benchmark code and revise some algorithms
Main development should now occur in bench.c, as these functions are
wholly identical besides tracking extra information. This info is useful
for determining the practical usage of these algorithms.

Safe Prime's still need to be implemented. The current roadblock is
generating random quadratic residues in m. This is a little tricky since
m = w * p, and w is not necessarily smooth. It may be required to alter
parameter generation to narrow this constraint.

Fast Prime Generation has a few hiccups that need to be ironed out. The
current 'fast' implementation is roughly the same speed as the naive
method, however it requires vastly less test-runs to identify primes.
This is a good sign, since expensive operation cost is addressed in the
faster variant.